pyproject.toml
[tool.poetry]
name = "audio-ml"
version = "0.1.0"
description = ""
authors = ["Prakrit Baruah <prakrit.baruah1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.39.0"
google-generativeai = "^0.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
